Refactor migrations to use Laravel conventions and add relationships

+++ database/migrations/2024_06_10_122942_add_names_roles_username_to_users_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('firstName');
             $table->string('lastName');
             $table->string('username');
-            $table->integer('roleNo');
+            $table->unsignedBigInteger('roleNo');
             $table->foreign('roleNo')->references('roleNo')->on('roles')->onUpdate('cascade')->onDelete('cascade');
         });
     }
